Fix NullPointerException issue after turn on airplane mode
Response may be cleared when SimRecord is reset,so check if it is
null before using response.

Fix issue with deletion of contacts with "=".

Fix issue for deletion of contacts with "=" in name,
as it would return 3 strings instead of 2 as expected.
Use overloaded split method to return 2 strings instead
to resolve the issue.

Bug:35026607
Change-Id: Ia20a646cb10bd15136de900df7177ca0bac4e465
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>